MPI for DMRG-SCF

+++ CheMPS2/CASSCFpt2.cpp
@@ -32,6 +32,7 @@
 #include "Lapack.h"
 #include "Cumulant.h"
 #include "CASPT2.h"
+#include "MPIchemps2.h"
 
 using std::string;
 using std::ifstream;
@@ -40,6 +41,12 @@ using std::endl;
 using std::max;
 
 double CheMPS2::CASSCF::caspt2( const int Nelectrons, const int TwoS, const int Irrep, ConvergenceScheme * OptScheme, const int rootNum, DMRGSCFoptions * scf_options, const double IPEA, const double IMAG, const bool PSEUDOCANONICAL ){
+
+   #ifdef CHEMPS2_MPI_COMPILATION
+      const bool am_i_master = ( MPIchemps2::mpi_rank() == MPI_CHEMPS2_MASTER );
+   #else
+      const bool am_i_master = true;
+   #endif
 
    const int num_elec = Nelectrons - 2 * iHandler->getNOCCsum();
    assert( num_elec >= 0 );
@@ -69,21 +76,32 @@ double CheMPS2::CASSCF::caspt2( const int Nelectrons, const int TwoS, const int 
    } else {
       if ( successful_solve == false ){
          assert( scf_options->getStoreUnitary() );
-         struct stat file_info;
-         const int file_stat = stat( (scf_options->getUnitaryStorageName()).c_str(), &file_info );
-         assert( file_stat == 0 );
-         unitary->loadU( scf_options->getUnitaryStorageName() );
+         if ( am_i_master ){
+            struct stat file_info;
+            const int file_stat = stat( (scf_options->getUnitaryStorageName()).c_str(), &file_info );
+            assert( file_stat == 0 );
+            unitary->loadU( scf_options->getUnitaryStorageName() );
+         }
+         #ifdef CHEMPS2_MPI_COMPILATION
+         unitary->broadcast( MPI_CHEMPS2_MASTER );
+         #endif
       }
    }
 
    // Fill active space Hamiltonian
-   Hamiltonian * HamAS = new Hamiltonian(nOrbDMRG, SymmInfo.getGroupNumber(), iHandler->getIrrepOfEachDMRGorbital());
-   Problem * Prob = new Problem(HamAS, TwoS, num_elec, Irrep);
-   Prob->SetupReorderD2h(); //Doesn't matter if the group isn't D2h, Prob checks it.
+   Hamiltonian * HamAS = new Hamiltonian( nOrbDMRG, SymmInfo.getGroupNumber(), iHandler->getIrrepOfEachDMRGorbital() );
+   Problem * Prob = new Problem( HamAS, TwoS, num_elec, Irrep );
+   Prob->SetupReorderD2h(); // Doesn't matter if the group isn't D2h, Prob checks it.
    buildTmatrix();
    buildQmatOCC();
    fillConstAndTmatDMRG( HamAS );
-   DMRGSCFrotations::rotate( VMAT_ORIG, HamAS->getVmat(), NULL, 'A', 'A', 'A', 'A', iHandler, unitary, mem1, mem2, work_mem_size, tmp_filename );
+   if ( am_i_master ){
+      DMRGSCFrotations::rotate( VMAT_ORIG, HamAS->getVmat(), NULL, 'A', 'A', 'A', 'A', iHandler, unitary, mem1, mem2, work_mem_size, tmp_filename );
+   }
+   #ifdef CHEMPS2_MPI_COMPILATION
+   HamAS->getVmat()->broadcast( MPI_CHEMPS2_MASTER );
+   #endif
+
    double E_CASSCF = 0.0;
    double * three_dm = new double[ tot_dmrg_power6 ];
    double * contract = new double[ tot_dmrg_power6 ];
@@ -91,37 +109,46 @@ double CheMPS2::CASSCF::caspt2( const int Nelectrons, const int TwoS, const int 
    // Solve the active space problem
    if (( OptScheme == NULL ) && ( rootNum == 1 )){ // Do FCI
 
-      const int nalpha = ( num_elec + TwoS ) / 2;
-      const int nbeta  = ( num_elec - TwoS ) / 2;
-      const double workmem = 1000.0; // 1GB
-      const int verbose = 2;
-      CheMPS2::FCI * theFCI = new CheMPS2::FCI( HamAS, nalpha, nbeta, Irrep, workmem, verbose );
-      double * inoutput = new double[ theFCI->getVecLength(0) ];
-      theFCI->ClearVector( theFCI->getVecLength(0), inoutput );
-      inoutput[ theFCI->LowestEnergyDeterminant() ] = 1.0;
-      E_CASSCF = theFCI->GSDavidson( inoutput );
-      theFCI->Fill2RDM( inoutput, DMRG2DM );                     // 2-RDM
-      theFCI->Fill3RDM( inoutput, three_dm );                    // 3-RDM
-      setDMRG1DM( num_elec, nOrbDMRG, DMRG1DM, DMRG2DM );        // 1-RDM
-      buildQmatACT();
-      construct_fock( theFmatrix, theTmatrix, theQmatOCC, theQmatACT, iHandler );
-      copy_active( theFmatrix, mem2, iHandler );                 // Fock
-      theFCI->Fock4RDM( inoutput, three_dm, mem2, contract );    // trace( Fock * 4-RDM )
-      delete theFCI;
-      delete [] inoutput;
+      if ( am_i_master ){
+         const int nalpha = ( num_elec + TwoS ) / 2;
+         const int nbeta  = ( num_elec - TwoS ) / 2;
+         const double workmem = 1000.0; // 1GB
+         const int verbose = 2;
+         CheMPS2::FCI * theFCI = new CheMPS2::FCI( HamAS, nalpha, nbeta, Irrep, workmem, verbose );
+         double * inoutput = new double[ theFCI->getVecLength(0) ];
+         theFCI->ClearVector( theFCI->getVecLength(0), inoutput );
+         inoutput[ theFCI->LowestEnergyDeterminant() ] = 1.0;
+         E_CASSCF = theFCI->GSDavidson( inoutput );
+         theFCI->Fill2RDM( inoutput, DMRG2DM );                     // 2-RDM
+         theFCI->Fill3RDM( inoutput, three_dm );                    // 3-RDM
+         setDMRG1DM( num_elec, nOrbDMRG, DMRG1DM, DMRG2DM );        // 1-RDM
+         buildQmatACT();
+         construct_fock( theFmatrix, theTmatrix, theQmatOCC, theQmatACT, iHandler );
+         copy_active( theFmatrix, mem2, iHandler );                 // Fock
+         theFCI->Fock4RDM( inoutput, three_dm, mem2, contract );    // trace( Fock * 4-RDM )
+         delete theFCI;
+         delete [] inoutput;
+      }
+      #ifdef CHEMPS2_MPI_COMPILATION
+      MPIchemps2::broadcast_array_double( &E_CASSCF, 1, MPI_CHEMPS2_MASTER );
+      MPIchemps2::broadcast_array_double(  DMRG2DM, dmrgsize_power4, MPI_CHEMPS2_MASTER );
+      MPIchemps2::broadcast_array_double( three_dm, tot_dmrg_power6, MPI_CHEMPS2_MASTER );
+      MPIchemps2::broadcast_array_double( contract, tot_dmrg_power6, MPI_CHEMPS2_MASTER );
+      setDMRG1DM( num_elec, nOrbDMRG, DMRG1DM, DMRG2DM );
+      #endif
 
    } else { // Do the DMRG sweeps
 
       assert( OptScheme != NULL );
       for ( int cnt = 0; cnt < dmrgsize_power4; cnt++ ){ DMRG2DM[ cnt ] = 0.0; } // Clear the 2-RDM
-      CheMPS2::DMRG * theDMRG = new DMRG(Prob, OptScheme);
-      for (int state = 0; state < rootNum; state++){
-         if (state > 0){ theDMRG->newExcitation( fabs( E_CASSCF ) ); }
+      CheMPS2::DMRG * theDMRG = new DMRG( Prob, OptScheme );
+      for ( int state = 0; state < rootNum; state++ ){
+         if ( state > 0 ){ theDMRG->newExcitation( fabs( E_CASSCF ) ); }
          E_CASSCF = theDMRG->Solve();
-         if ((state == 0) && (rootNum > 1)){ theDMRG->activateExcitations( rootNum-1 ); }
+         if (( state == 0 ) && ( rootNum > 1 )){ theDMRG->activateExcitations( rootNum - 1 ); }
       }
       theDMRG->calc_rdms_and_correlations( true );
-      copy2DMover( theDMRG->get2DM(), nOrbDMRG, DMRG2DM  ); // 2-RDM
+      copy2DMover( theDMRG->get2DM(), nOrbDMRG, DMRG2DM ); // 2-RDM
       setDMRG1DM( num_elec, nOrbDMRG, DMRG1DM, DMRG2DM ); // 1-RDM
       buildQmatACT();
       construct_fock( theFmatrix, theTmatrix, theQmatOCC, theQmatACT, iHandler );
@@ -159,7 +186,7 @@ double CheMPS2::CASSCF::caspt2( const int Nelectrons, const int TwoS, const int 
    delete HamAS;
 
    if ( PSEUDOCANONICAL == false ){
-      cout << "CASPT2 : Deviation from pseudocanonical = " << deviation_from_blockdiag( theFmatrix, iHandler ) << endl;
+      if ( am_i_master ){ cout << "CASPT2 : Deviation from pseudocanonical = " << deviation_from_blockdiag( theFmatrix, iHandler ) << endl; }
       block_diagonalize( 'O', theFmatrix, unitary, mem1, mem2, iHandler, false, NULL, NULL, NULL );
       block_diagonalize( 'A', theFmatrix, unitary, mem1, mem2, iHandler, false, DMRG2DM, three_dm, contract ); // 2-RDM, 3-RDM, and trace( Fock * cu(4)-4-RDM )
       block_diagonalize( 'V', theFmatrix, unitary, mem1, mem2, iHandler, false, NULL, NULL, NULL );
@@ -170,19 +197,27 @@ double CheMPS2::CASSCF::caspt2( const int Nelectrons, const int TwoS, const int 
       construct_fock( theFmatrix, theTmatrix, theQmatOCC, theQmatACT, iHandler ); // Fock
    }
 
-   //Calculate the matrix elements needed to calculate the CASPT2 V-vector
-   DMRGSCFrotations::rotate( VMAT_ORIG, NULL, theRotatedTEI, 'C', 'C', 'F', 'F', iHandler, unitary, mem1, mem2, work_mem_size, tmp_filename );
-   DMRGSCFrotations::rotate( VMAT_ORIG, NULL, theRotatedTEI, 'C', 'V', 'C', 'V', iHandler, unitary, mem1, mem2, work_mem_size, tmp_filename );
-   delete_file( tmp_filename );
+   // Calculate the matrix elements needed to calculate the CASPT2 V-vector
+   if ( am_i_master ){
+      DMRGSCFrotations::rotate( VMAT_ORIG, NULL, theRotatedTEI, 'C', 'C', 'F', 'F', iHandler, unitary, mem1, mem2, work_mem_size, tmp_filename );
+      DMRGSCFrotations::rotate( VMAT_ORIG, NULL, theRotatedTEI, 'C', 'V', 'C', 'V', iHandler, unitary, mem1, mem2, work_mem_size, tmp_filename );
+      delete_file( tmp_filename );
+   }
 
    delete [] mem1;
    delete [] mem2;
 
-   cout << "CASPT2 : Deviation from pseudocanonical = " << deviation_from_blockdiag( theFmatrix, iHandler ) << endl;
-   CheMPS2::CASPT2 * myCASPT2 = new CheMPS2::CASPT2( iHandler, theRotatedTEI, theTmatrix, theFmatrix, DMRG1DM, DMRG2DM, three_dm, contract, IPEA );
-   const double E_CASPT2 = myCASPT2->solve( IMAG );
+   double E_CASPT2 = 0.0;
+   if ( am_i_master ){
+      cout << "CASPT2 : Deviation from pseudocanonical = " << deviation_from_blockdiag( theFmatrix, iHandler ) << endl;
+      CheMPS2::CASPT2 * myCASPT2 = new CheMPS2::CASPT2( iHandler, theRotatedTEI, theTmatrix, theFmatrix, DMRG1DM, DMRG2DM, three_dm, contract, IPEA );
+      E_CASPT2 = myCASPT2->solve( IMAG );
+      delete myCASPT2;
+   }
+   #ifdef CHEMPS2_MPI_COMPILATION
+   MPIchemps2::broadcast_array_double( &E_CASPT2, 1, MPI_CHEMPS2_MASTER );
+   #endif
 
-   delete myCASPT2;
    delete [] three_dm;
    delete [] contract;
